% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcards.R
\name{jolla}
\alias{jolla}
\title{Jolla website template.}
\usage{
jolla(css = NULL, includes = NULL, self_contained = TRUE)
}
\arguments{
\item{css}{One or more css files to include}

\item{includes}{Named list of additional content to include within the
document (typically created using the \code{\link[rmarkdown]{includes}} function).}

\item{self_contained}{Produce a standalone HTML file with no external
dependencies, using data: URIs to incorporate the contents of linked
scripts, stylesheets, images, and videos. Note that even for self contained
documents MathJax is still loaded externally (this is necessary because of
its size).}
}
\description{
Jolla website template.
}
