% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcards.R
\name{jolla_blue}
\alias{jolla_blue}
\title{Jolla Blue website template.}
\usage{
jolla_blue(css = NULL, includes = NULL)
}
\arguments{
\item{css}{One or more css files to include}

\item{includes}{Named list of additional content to include within the
document (typically created using the \code{\link[rmarkdown]{includes}} function).}
}
\description{
Jolla Blue website template.
}
