% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcards.R
\name{solana}
\alias{solana}
\title{Solana website template.}
\usage{
solana(css = NULL, includes = NULL)
}
\arguments{
\item{css}{CSS and/or Sass files to include. Files with an extension of .sass
or .scss are compiled to CSS via \code{sass::sass()}. Also, if \code{theme} is a
\code{\link[bslib:bs_theme]{bslib::bs_theme()}} object, Sass code may reference the relevant Bootstrap
Sass variables, functions, mixins, etc.}

\item{includes}{Named list of additional content to include within the
document (typically created using the \code{\link[rmarkdown]{includes}} function).}
}
\description{
Solana website template.
}
